Share sort_values between Series and DataFrame.
